Refactoring: extract AbstractProposalComputer
makes it easier to add more implementations of proposal computers
